lead time: add flag to use a fallback for the start time value
Some issues are never moved to the "In Progress" column in Zenhub, for
all kinds of different reasons. Those would usually be skipped in the
lead time statistics but in case a user wants to include them, the new
flag lets them state this expectation. In this case the start time is
pulled from the issue creation time value.

Since this may lead to a significant different in lead times, a
warning is printed.